requirements.txt
sondehub
python-dateutil
requests
shapely
pytz